refactor: fix auth service types and rename repository files
