+++ lib/domain/usecases/auth/get_user.dart
@@ -1,6 +1,6 @@
 import 'package:dartz/dartz.dart';
 import 'package:tech_challenge_3/core/usecase/usecase.dart';
-import 'package:tech_challenge_3/domain/repository/auth.dart';
+import 'package:tech_challenge_3/domain/repository/auth_repository.dart';
 import 'package:tech_challenge_3/service_locator.dart';
 
 class GetUserUseCase implements UseCase<Either, dynamic> {
