+++ lib/domain/usecases/auth/is_logged_in.dart
@@ -1,5 +1,5 @@
 import 'package:tech_challenge_3/core/usecase/usecase.dart';
-import 'package:tech_challenge_3/domain/repository/auth.dart';
+import 'package:tech_challenge_3/domain/repository/auth_repository.dart';
 import 'package:tech_challenge_3/service_locator.dart';
 
 class IsLoggedInUseCase implements UseCase<bool, dynamic> {
