+++ lib/domain/usecases/transactions/create_transaction.dart
@@ -1,8 +1,8 @@
 import 'package:dartz/dartz.dart';
 import 'package:tech_challenge_3/core/usecase/usecase.dart';
 import 'package:tech_challenge_3/data/models/transaction_create_dto.dart';
-import 'package:tech_challenge_3/domain/repository/auth.dart';
-import 'package:tech_challenge_3/domain/repository/transactions.dart';
+import 'package:tech_challenge_3/domain/repository/auth_repository.dart';
+import 'package:tech_challenge_3/domain/repository/transactions_repository.dart';
 import 'package:tech_challenge_3/service_locator.dart';
 
 class CreateTransactionUseCase
