+++ lib/domain/usecases/transactions/delete_transaction.dart
@@ -1,6 +1,6 @@
 import 'package:dartz/dartz.dart';
 import 'package:tech_challenge_3/core/usecase/usecase.dart';
-import 'package:tech_challenge_3/domain/repository/transactions.dart';
+import 'package:tech_challenge_3/domain/repository/transactions_repository.dart';
 import 'package:tech_challenge_3/service_locator.dart';
 
 class DeleteTransactionUseCase implements UseCase<Either, String> {
