+++ lib/domain/usecases/transactions/update_transaction.dart
@@ -1,7 +1,7 @@
 import 'package:dartz/dartz.dart';
 import 'package:tech_challenge_3/core/usecase/usecase.dart';
 import 'package:tech_challenge_3/data/models/transaction_update_dto.dart';
-import 'package:tech_challenge_3/domain/repository/transactions.dart';
+import 'package:tech_challenge_3/domain/repository/transactions_repository.dart';
 import 'package:tech_challenge_3/service_locator.dart';
 
 class UpdateTransactionUseCase
