+++ lib/domain/usecases/transactions/upload_transaction_attachment.dart
@@ -1,7 +1,7 @@
 import 'package:dartz/dartz.dart';
 import 'package:tech_challenge_3/core/usecase/usecase.dart';
 import 'package:tech_challenge_3/data/models/upload_attachment_params.dart';
-import 'package:tech_challenge_3/domain/repository/transactions.dart';
+import 'package:tech_challenge_3/domain/repository/transactions_repository.dart';
 import 'package:tech_challenge_3/service_locator.dart';
 
 class UploadTransactionAttachmentUseCase
