+++ lib/presentation/transactions/bloc/transactions_display_cubit.dart
@@ -41,13 +41,10 @@ class TransactionsDisplayCubit extends Cubit<TransactionsDisplayState> {
   Future<void> fetchTransactions() async {
     emit(TransactionsDisplayLoading());
     final result = await _getTransactionsUseCase.call();
-    result.fold(
-      (error) => emit(TransactionsDisplayError(errorMessage: error)),
-      (transactions) {
-        final txList = List<TransactionEntity>.from(transactions);
-        txList.sort((a, b) => b.date.compareTo(a.date));
-        emit(TransactionsDisplayLoaded(transactions: txList));
-      },
-    );
+    result.fold((transactions) {
+      final txList = List<TransactionEntity>.from(transactions);
+      txList.sort((a, b) => b.date.compareTo(a.date));
+      emit(TransactionsDisplayLoaded(transactions: txList));
+    }, (error) => emit(TransactionsDisplayError(errorMessage: error)));
   }
 }
